+++ backend/src/application/use_cases/tasks/tasks_use_cases.py
@@ -1,9 +1,8 @@
 from datetime import datetime, timezone
 
 from sqlalchemy import select
-from sqlalchemy.orm import selectinload
 
-from backend.src.infrastructure.database.models import ExternalTask, TaskType, User
+from backend.src.infrastructure.database.models import ExternalTask
 from backend.src.infrastructure.database.repositories import (
     ExternalSystemRepository,
     ExternalTaskRepository,
@@ -14,11 +13,7 @@ from backend.src.presentation.dto import (
     ExternalTaskCreateRequest,
     ExternalTaskResponse,
     ExternalTaskUpdateRequest,
-    TaskTypeRequest,
-    TaskTypeResponse,
-    TaskTypeUpdateRequest,
 )
-from backend.src.presentation.dto.converters import task_type_to_response
 
 
 class TasksUseCase:
@@ -140,108 +135,4 @@ class TasksUseCase:
         except Exception as e:
             raise InternalServerError(
                 f"Failed to delete task: {str(e)}", {"task_id": task_id, "error": str(e)}
-            )
-
-    async def get_types(self, user_id: int) -> list[TaskTypeResponse]:
-        """Get all task types for a user."""
-        try:
-            result = await self.user_repository.session.execute(
-                select(User).options(selectinload(User.task_types)).where(User.id == user_id)
-            )
-            user = result.scalar_one_or_none()
-
-            if not user:
-                return []
-
-            return [task_type_to_response(task_type) for task_type in user.task_types]
-        except Exception as e:
-            raise InternalServerError(f"Failed to get task types: {str(e)}", {"user_id": user_id})
-
-    async def create_type(self, user_id: int, data: TaskTypeRequest) -> TaskTypeResponse:
-        """Create a new task type for a user."""
-        try:
-            result = await self.user_repository.session.execute(
-                select(User).options(selectinload(User.task_types)).where(User.id == user_id)
-            )
-            user = result.scalar_one_or_none()
-
-            if not user:
-                raise NotFoundError("User", details={"user_id": user_id})
-
-            task_type = TaskType(
-                user_id=user.id,
-                title=data.title,
-                color=data.color,
-            )
-            user.task_types.append(task_type)
-            await self.user_repository.session.commit()
-            return task_type_to_response(task_type)
-        except NotFoundError:
-            raise
-        except Exception as e:
-            raise InternalServerError(f"Failed to create task type: {str(e)}", {"user_id": user_id})
-
-    async def update_type(
-        self, task_type_id: int, data: TaskTypeUpdateRequest, user_id: int
-    ) -> TaskTypeResponse:
-        """Update an existing task type."""
-        try:
-            result = await self.user_repository.session.execute(
-                select(User)
-                .options(selectinload(User.task_types))
-                .join(TaskType, User.id == TaskType.user_id)
-                .where(TaskType.id == task_type_id)
-                .where(User.id == user_id)
-            )
-            user = result.scalar_one_or_none()
-
-            if not user:
-                raise NotFoundError("TaskType", task_type_id)
-
-            task_type = next((tt for tt in user.task_types if tt.id == task_type_id), None)
-            if not task_type:
-                raise NotFoundError("TaskType", task_type_id)
-
-            if data.title is not None:
-                task_type.title = data.title
-            if data.color is not None:
-                task_type.color = data.color
-            if data.is_active is not None:
-                task_type.is_active = data.is_active
-
-            await self.user_repository.session.commit()
-            return task_type_to_response(task_type)
-        except NotFoundError:
-            raise
-        except Exception as e:
-            raise InternalServerError(
-                f"Failed to update task type: {str(e)}", {"task_type_id": task_type_id}
-            )
-
-    async def delete_type(self, task_type_id: int, user_id: int) -> None:
-        """Delete a task type."""
-        try:
-            result = await self.user_repository.session.execute(
-                select(User)
-                .options(selectinload(User.task_types))
-                .join(TaskType, User.id == TaskType.user_id)
-                .where(TaskType.id == task_type_id)
-                .where(User.id == user_id)
-            )
-            user = result.scalar_one_or_none()
-
-            if not user:
-                raise NotFoundError("TaskType", task_type_id)
-
-            task_type = next((tt for tt in user.task_types if tt.id == task_type_id), None)
-            if not task_type:
-                raise NotFoundError("TaskType", task_type_id)
-
-            user.task_types.remove(task_type)
-            await self.user_repository.session.commit()
-        except NotFoundError:
-            raise
-        except Exception as e:
-            raise InternalServerError(
-                f"Failed to delete task type: {str(e)}", {"task_type_id": task_type_id}
             )
