+++ backend/src/presentation/controllers/profile/profile_controller.py
@@ -2,8 +2,10 @@ from dishka import FromDishka
 from dishka.integrations.litestar import inject
 from litestar import Controller, Request, delete, get, patch, post
 
+from backend.src.application.use_cases.ai.ai_preferences_use_cases import AIPreferencesUseCases
+from backend.src.application.use_cases.auth.user_use_cases import UserUseCases
 from backend.src.application.use_cases.settings.settings_use_cases import SettingsUseCases
-from backend.src.application.use_cases.tasks.tasks_use_cases import TasksUseCase
+from backend.src.application.use_cases.tasks.task_type_use_cases import TaskTypeUseCases
 from backend.src.presentation.dto.auth import UserUpdateRequest, UserUpdateRequestDTO
 from backend.src.presentation.dto.settings import (
     AIPreferencesUpdateRequest,
@@ -27,11 +29,11 @@ class ProfileController(Controller):
         self,
         data: UserUpdateRequest,
         request: Request,
-        settings_use_cases: FromDishka[SettingsUseCases],
+        user_use_cases: FromDishka[UserUseCases],
     ) -> SuccessResponse:
         """Update user profile information."""
         user_id = request.user.id
-        updated_user = await settings_use_cases.update_user(user_id=user_id, data=data)
+        updated_user = await user_use_cases.update(user_id=user_id, data=data)
         return SuccessResponse(message="User updated successfully", data=updated_user)
 
     @patch("/ai-preferences", dto=AIPreferencesUpdateRequestDTO, return_dto=SuccessResponseDTO)
@@ -40,11 +42,11 @@ class ProfileController(Controller):
         self,
         data: AIPreferencesUpdateRequest,
         request: Request,
-        settings_use_cases: FromDishka[SettingsUseCases],
+        aI_preferences_use_cases: FromDishka[AIPreferencesUseCases],
     ) -> SuccessResponse:
         """Update user's AI preferences (provider selection, auto-processing)."""
         user_id = request.user.id
-        updated_preferences = await settings_use_cases.update_user_ai_preferences(
+        updated_preferences = await aI_preferences_use_cases.update_user_preferences(
             user_id=user_id, data=data
         )
         return SuccessResponse(
@@ -54,11 +56,11 @@ class ProfileController(Controller):
     @get("/ai-preferences/providers", return_dto=SuccessResponseDTO)
     @inject
     async def get_available_ai_providers(
-        self, request: Request, settings_use_cases: FromDishka[SettingsUseCases]
+        self, request: Request, ai_preferences_use_cases: FromDishka[AIPreferencesUseCases]
     ) -> SuccessResponse:
         """Get available AI providers for user selection."""
         user_id = request.user.id
-        providers = await settings_use_cases.get_active_ai_providers(user_id=user_id)
+        providers = await ai_preferences_use_cases.get_active_providers(user_id=user_id)
         return SuccessResponse(
             message="Available AI providers retrieved successfully", data=providers
         )
@@ -69,11 +71,11 @@ class ProfileController(Controller):
         self,
         request: Request,
         settings_use_cases: FromDishka[SettingsUseCases],
-        task_types_use_case: FromDishka[TasksUseCase],
+        task_types_use_case: FromDishka[TaskTypeUseCases],
     ) -> SuccessResponse:
         """Get user's task types."""
         user_id = request.user.id
-        task_types = await task_types_use_case.get_types(user_id=user_id)
+        task_types = await task_types_use_case.get_many(user_id=user_id)
         return SuccessResponse(message="Task types retrieved successfully", data=task_types)
 
     @post("/task-types", dto=TaskTypeRequestDTO, return_dto=SuccessResponseDTO)
@@ -83,11 +85,11 @@ class ProfileController(Controller):
         data: TaskTypeRequest,
         request: Request,
         settings_use_cases: FromDishka[SettingsUseCases],
-        task_types_use_case: FromDishka[TasksUseCase],
+        task_types_use_case: FromDishka[TaskTypeUseCases],
     ) -> SuccessResponse:
         """Create new task type for user."""
         user_id = request.user.id
-        task_type = await task_types_use_case.create_type(user_id=user_id, data=data)
+        task_type = await task_types_use_case.create(user_id=user_id, data=data)
         return SuccessResponse(message="Task type created successfully", data=task_type)
 
     @patch(
@@ -102,11 +104,11 @@ class ProfileController(Controller):
         data: TaskTypeUpdateRequest,
         request: Request,
         settings_use_cases: FromDishka[SettingsUseCases],
-        task_types_use_case: FromDishka[TasksUseCase],
+        task_types_use_case: FromDishka[TaskTypeUseCases],
     ) -> SuccessResponse:
         """Update user's task type."""
         user_id = request.user.id
-        task_type = await task_types_use_case.update_type(
+        task_type = await task_types_use_case.update(
             task_type_id=task_type_id, data=data, user_id=user_id
         )
         return SuccessResponse(message="Task type updated successfully", data=task_type)
@@ -117,11 +119,11 @@ class ProfileController(Controller):
         self,
         task_type_id: int,
         request: Request,
-        settings_use_cases: FromDishka[SettingsUseCases],
+        task_types_use_case: FromDishka[TaskTypeUseCases],
     ) -> None:
         """Delete user's task type."""
         user_id = request.user.id
-        await settings_use_cases.delete_task_type(task_type_id=task_type_id, user_id=user_id)
+        await task_types_use_case.delete(task_type_id=task_type_id, user_id=user_id)
 
     @patch(
         "/ai-preferences/providers/{ai_provider_id:int}",
@@ -135,10 +137,11 @@ class ProfileController(Controller):
         request: Request,
         data: AIProviderUpdateRequest,
         settings_use_cases: FromDishka[SettingsUseCases],
+        ai_preferences_use_case: FromDishka[AIPreferencesUseCases],
     ) -> SuccessResponse:
         """Update AI provider configuration (admin only)."""
         user_id = request.user.id
-        updated_provider = await settings_use_cases.update_ai_provider(
+        updated_provider = await ai_preferences_use_case.update_provider(
             ai_provider_id=ai_provider_id, data=data, user_id=user_id
         )
         return SuccessResponse(message="AI provider updated successfully", data=updated_provider)
